requirements.txt
Django==6.0.1
python-dotenv==1.2.1
groq==1.0.0
PyPDF2==3.0.1
python-docx>=1.1.0
openpyxl>=3.1.2
Pillow>=10.0.0